Change k.Graph's data format to datum style.
This means that instead of an array of x/y pairs for each line on the
graph (which duplicates data like the created times), there is an array
of objects which represent a moment in time, and all the values at that
time.

This assumes that each point in time will have a value for all of the
data points, or at least most of them. Rickshaw does not make this
assumption, but k.Graph now does, mainly because it has held true in all
the data we have used so far.

This makes progress towards completing bug 865378.