[Bug 865378] Add support for multiple Y axises.
This uses "axis groups". Each axis group's maximum is calculated, and
then the entire group is scaled so that the maximum is 1.0. Then the
axises and the hover tool tips are adjusted to show the correct numbers.
The scaling is adjusted as the zoom is changed and as acises are toggled
on and off.

This is kind of hacky, and the more I do this, the more I want to just
use d3 directly.