Changed ChoiceField to TypedChoiceField
Significance was an unicode string from the fields. `TypedChoiceField`
will force `significance` to be converted to an integer.

Without this, the value of `significance` will be wrong until it is
retrieved from the database, as I have encountered when trying to
compare the value of `signficance` in `wiki.review_revision`