[bug 892610] Patch TestWikiVideo with mediocrity
This changes the test so it accepts the previous output I was getting
and the new output I'm getting now. I suspect the difference in my
system is that I updated from libxml2 2.8 to 2.9.1 and in doing that
I get slightly different output.

This is goofy, but I think both outputs are "good enough" and I think
Ricky said the video stuff in the wiki isn't long for this world anyhow.