[bug 889833] Fix excerpts for contributor forum searches
This re-adds excerpts for contributor forum searches.

Further, it fixes a bug where we were setting number_of_fragments to 0
which causes ES to highlight the entire content and ignore fragment_size.
If we set it to 1, then it picks one fragment and limits it to the
fragment_size.

That might cause other issues, but it's the right thing to do otherwise
excerpts can get huge especially for support and contributor forum results.