[bug 904443] Fix highlighting
This reinstates the "show the whole thing" highlighting.

Because we're showing the whole field highlighted, "fragment_size"
has no effect, so I nixed that.

Additionally, there's no "discussion_content" field so I nixed
that, too.

On top of that, to prevent me from trying to fix this again in the
future, I updated the comment so that it correctly represents
what should be happening in the code.