[Bug 906992] Set up multiple ES index framework.
Now settings.py can contain multiple indexes in ES_INDEXES and/or
ES_WRITE_INDEXES, and models can choose which kind of index to use. The
admin and management commands have been updated, and I updated search
and other parts of the site that use elastic to account for multiple
indexes, where needed.