[Bug 914268] Suggests username during Persona signup
Added suggested username to persona signup.

One interesting case for this method is if the user sign up with
something like: john1@gmail.com and john1 is taken. The suggestion will
be john1[0-9]+

However, I think this is acceptable.

Another note is right now the suggested username gets dropped as a value
of the form prefilled.