Use dennis_shim.py for 10x speedup
Turns out running dennis in ./manage.py is uber-slow because of all the
django/whatever stuff. Running dennis directly with a shim that sets up
the sys.path correctly is *so much faster*.

Also, this updates dennis to v0.3.8 which picks up a fix for false
positives in mismatched variables error detection.