[bug 932301] Collapse whitespace in db strings
Tower collapses whitespace in a string before passing it to gettext to
be translate. The problem here is that we weren't collapsing whitespace
in db strings. So when tower goes to do the gettext thing, it would
collapse the whitespace and then the string wouldn't match a msgid and
then it wouldn't get translated.

This works fine for single line strings because they don't have
whitespace to collapse. Looks like all the db strings were one-liners so
we never saw this problem until now.

This fixes it by calling strip_whitespace on the strings coming out of
the db so they match what tower is doing.