Add setting to skip migrations when building test db
We periodically add data via data migrations. However with test-utils,
the first time you run tests that depend on that new data, they work but
subsequent tests will fail.

However, this is the same situation we had with schematic. We use model
makers to generate data we need for tests.

Ergo, there's no reason to run the south migrations to create the test
db.

This shuts that off.