[bug 943464] Add reconciling to reindexing
Our existing reindexing code would update the index with db contents,
but it never checked the index to see if the index had documents in it
that weren't in the db.

This adds reconcilation code that removes things from the index that
shouldn't be there.

This also tweaks the admin page for index maintenance and adds an
outstanding records section making it easier to see what's going on and
also easier to see what happened.