[bug 947206] Change log_error to log_exception
Right now one of the strings is kicking up an error, but we have no idea
which string it is because there's no traceback. This changes the
log_error to a log_exception which will give us a traceback which
hopefully will make it to Sentry and then we should be able to track
down the problem string.