Fix false positive linting errors regarding %
These strings use %, but the problem is that translations of these
strings could use % such that dennis thinks there's a python variable
there which kicks up an "invalid variables" error which prevents the .po
file from being compiled to an .mo file.

This fixes the strings by adding a dennis-ignore note to the extracted
comments. Dennis will see the note and skip running the lint rules on
those strings.