[Bug 1074959] Add metadata API.
This isn't as nice as I wanted. The metadata field of questions is a
list of objects in the API. Semantically it should be a dict, but I
couldn't get DRF to play nice with writing in that format. There is some
support for writing directly to the object with POST and PATCH, but it's
kind of derpy, so I also added endpoints for `set_metadata` and
`delete_metadata` that work intuitively for single items.