[bug 1056298] Fix createsuperuser
This adds a patch call to manage.py. The problem is that various
Django-specific subcommands at some point import Jingo which eventually
results in a circular import. Adding the patch call here guarantees
that Jingo and all that stuff is imported before we import the things we
need to import to execute the subcommand.