[Bug 1083417] Add django-action-stream.
This uses django-action-stream instead of django-notifications like
specified in the bug because django-notifications is actually a UI-like
Django app that provides more than we need. Plus, it's app name is
"notifications", and we already have one of those so it was annoying.
django-notifications is based on django-activity-stream, so the data
model that was discussed previously still applies.