[bug 1148082] Fix --no-skip
nose has a skip plugin which enhances SkipTest raising beyond our
wildest imaginations. However, there are certain occasions where we want
skipped things to not skip with reckless abandon but instead don the
malevolent cloak of errorhood.

nose has a --no-skip argument for exactly this sort of sordid thing, but
it's totally busted and doesn't work.

This changes the code around so that it works correctly.