Get serious with str/unicode in in_review.py
Bugzilla returns JSON with utf-8 encoded strings which get converted to
unicode. So the bugzilla-using code is always dealing with unicode.

git data is pulled from the command line as byte strings. As such, it's
utf-8 encoded strings.

This commit straightens out the code so that we have everything as
unicode types and then explicitly encode as utf-8 when printing to the
command line.