Remove rabbitmq queue size measuring code
The -stage environment is somehow different than the -dev, -prod and all
developer environments such that the code we have for measuring the
rabbitmq queue size is erroring out.

We were faced with two options:

1. remove the problematic code and thus unblock the Django 1.7 upgrade

2. try to fix code that we know is fragile because it uses private
   methods which is broken on only one environment for reasons we
   don't understand  and in a way we can't easily observe the specifics
   of or test fixes for

We chose the first option because the second option is clearly a
slippery slope to giving up and becoming a potato farmer.  Also, yay
for -stage environments that differ in mysterious ways from -prod.