[bug 1164205] Added api docs for search suggest api
I have no idea if this is the format we want to go with, but it is some
documentation.

Maintaining it will be painful. It's probably better for us to figure
out a way to autogenerate the documentation or somehow move the
documentation to the code. That work is covered in bug #1164247.