[bug 1164565] Fix max_questions/documents in search suggest api
This fixes the code to limit the results returned by max_questions and
max_documents.

Further, it adds tests for this behavior so that it's less likely to
break again in the future when I'm making changes.

It also tweaks _make_question() and _make_document() to have a timestamp
in the title to prevent slug collisions.