[bug 702276] Upgrade lxml to 3.4.4 and pyquery to 1.2.9
This upgrades lxml to 3.4.4. In order to do that, I also had to upgrade
pyquery and pull in cssselect.

I think there's a bug in the cssselect handling for the :first
pseudo-element. That breaks a few of our tests. It kind of looks like
both pyquery and cssselect are dead-ish projects and aren't being
actively maintained, so at some point we might want to look at replacing
pyquery with something else like BeautifulSoup4.

I talked with Rehan about this a bit and decided that since it's just a
few tests that are busted, I'd just subclass PyQuery and provide a
.first() method.

The other thing I discovered when doing this is that the messages page
has changed significantly since 2011 and one of the tests not only fails
now, but it doesn't appear to be a valid test anyhow, so I removed it.