[bug 1116489] Tweak ImportError message in manage.py
If they have a virtual environment activated, then the problem is
probably either they haven't installed the requirements or they haven't
updated them in a while.

If they don't have a virtual environment actiivated, then they need to
activate it.

This code change attempts to distinguish between the two by looking for
VIRTUAL_ENV in the environment.