[bug 1212476] Refactor simple search to share code
* moves some things over to the SimpleSearchForm
* extracts S building code into simple_search.py:generate_simple_search
  so that it can be re-used between simple_search and
  opensearch_suggestions
* redoes opensearch_suggestions to use the new generate_simple_search

One thing this does is changes how opensearch_suggestions works.
Previously, it would do a search and show the top 5 kb articles followed
by the top 5 questions. That's what simple search used to do eons ago.

This changes that so that it does what simple search does and shows the
top 10 mixed results. This reduces the number of different kinds of
search we're doing by 1 and reduces the likelihood that
opensearch_suggestions gets totally out of sync with the other searches
like it did.