[bug 917754] Rewrite bookkeeping for indexing from admin
This redoes the bookkeeping for indexing from the admin. It fleshes
out the Record model to allow us to group records in batches. It
cleans up the admin page so it's easier to see outstanding records and
recent records and the data is more informative. It no longer uses
redis to keep track of the overall status. Plus it should be better
with celery since we're passing class paths around and not actual
classes.

Hooray!