[bug 1225557] Remove db/index reconcile code
Mike and I talked about this for a while and he didn't know this code
existed and we're pretty sure there's been no use of it for over a
year. Given that, I claim the problem this code was written to solve
is gone or something along those lines. Further this code needs to be
hand tested since there are no unit tests for it.

Given that, I'm removing it which will make it easier to overhaul
elasticsearch stuff.