[fix bug 1381436] Replace ES facets with DB queries.
ElasticSearch v2 has deprecated facets in favor of aggregations. To upgrade to
aggregations we need to update elastic search python libraries. This is a
combination of ES and Database queries to serve as a workaround until we move
codebase to v2.

Since the DB is quite slower compared to ES, the results get cached for 15
minutes.