[fix issue #3179] Switch product icons from sprite to real product icon.
Kitsune Products featured code to auto-generate a sprite with all product icons
on every product save. This code broken when we switched to the AWS S3 backed
user media backend.

Instead of fixing this code, we opted for removing the sprite functionality
altogether because:

  - it was not used in all places (i.e. questions used a different manually
    created sprite)

  - it raised serious concerns on how we can atomically update the auto
    generated sprite on S3 while avoiding errors.

  - The whole implementation was a hack which saved files without using
    default_storage engine.

Now product images are directly used on desktop and mobile and as a bonus the
same images -which can be managed through the admin interface- are used in both
product listing and AAQ.

Partially fixes bug 1196475 too.