Include service and action parameter for both links to FxA as well as the form. (#1308)
These parameters were hardcoded into the form but weren't being
inserted into the alternative links for signing in / signing up.

This patch makes it so that we provide these parameters as part
of the state, rather than hardcoding them into the form. This
makes us more consistent with the other parameters.