Add a dialog to the switching-devices form wizard for calendar reminders.
The dialog also allows the user to copy the download URL to their clipboard.

The dialog is not yet wired up to anything, but can be invoked by calling
`openReminderDialog()` on the setup-device-step element.

In this patch, I also needed to update some pre-existing tests to clean up
their spies/stubs so that the form-wizard-reminder-dialog-tests.js could
run correctly.

Fixes mozilla/sumo#1494.