Update copy for switching devices wizard.
This is the latest iteration for the copy in the switching devices wizard. This
also makes the ICS file download filename something that can be localized to
something meaningful.

The link in the calendar invite is now wrapped in an anchor with meaningful
text for Google Calendar, and just the URL for iCal (since iCal clients might
not display the anchor in a friendly way).

Fixes mozilla/sumo#1529.