Don't let users (or tests!) spam Basket so much.
mozilla/sumo#1562 was originally about disabling the submit button
sooner in the form submission process so that quick clicks on the
submission button wouldn't result in multiple XHRs while we waited
for Basket to respond.

While writing a test for that, I also noticed that our mocha tests
were actually sending real XHRs to the Basket endpoint, both when
running in automation and when running locally. Hilarious! This
patch fixes that too.

Fixes mozilla/sumo#1562