Device migration 3.0 (#5771)
* Add an email field and calendar reminder button to Step 3 of the switching-devices wizard.

Fixes mozilla/sumo#1505.

* Attempt to prefill email address in switching devices wizard step 3

This uses session storage to store the email address inputted by the
user in Step 1, and retrieves it after the FxA redirect in Step 3.

In the event that the user never went through Step 1, this is a no-op.

Fixes mozilla/sumo#1520.

* Update copy for switching devices wizard.

This is the latest iteration for the copy in the switching devices wizard. This
also makes the ICS file download filename something that can be localized to
something meaningful.

The link in the calendar invite is now wrapped in an anchor with meaningful
text for Google Calendar, and just the URL for iCal (since iCal clients might
not display the anchor in a friendly way).

Fixes mozilla/sumo#1529.

* Remove Outlook.com from switching-devices wizard.

Outlook.com has issues with deeplinking into the event editor. While
we work with the Outlook.com team to see if we can get that sorted,
we're going to remove Outlook.com as one of the calendar options.

Fixes mozilla/sumo#1527.

* Add more GA events for the switching-devices reminder dialog.

This adds GA events for when the reminder dialog is closed, regardless
of how it's closed (in case, for example, it's closed by the keyboard),
as well as an event that describes which kind of calendar was selected
from the dropdown.

Closes mozilla/sumo#1528.

* Submit request for switching-devices email in the document language.

The email request to Basket was originally using navigator.language,
but this doesn't take into account users that might be viewing the
switching-devices page in a language that doesn't match the browser
default.

This patch makes it so that we can pull the language code off of the
document element, and use that when we submit the newsletter request.

If for some reason, the language attribute isn't set, we fallback
to using navigator.language.

* Don't let users (or tests!) spam Basket so much.

mozilla/sumo#1562 was originally about disabling the submit button
sooner in the form submission process so that quick clicks on the
submission button wouldn't result in multiple XHRs while we waited
for Basket to respond.

While writing a test for that, I also noticed that our mocha tests
were actually sending real XHRs to the Basket endpoint, both when
running in automation and when running locally. Hilarious! This
patch fixes that too.

Fixes mozilla/sumo#1562

* Update download links created for calendar events

This uses the updated mzl.la link for the calendar events so that
they can be properly attributed.

Fixes mozilla/sumo#1503

---------

Co-authored-by: Mike Conley <mike.d.conley@gmail.com>
Co-authored-by: Mike Conley <mconley@mozilla.com>