Locale switcher from parent (#6708)
not in the default locale

This covers a case where the user is on a document that
is not in the default locale (in our case en-US) and
is not a translation.
If the user tries to use the locale switcher to go to
English, since there is no parent as we are not on a
translation, the user should get a 404.