Add SUMO rules mentions & Improve AAQ step 3
Add SUMO rules mentions when creating/editing a question, an answer, a forum thread, a forum post, a KB forum thread or a KB forum post.
In AAQ step 3, replace "Subject:" with "Summarize your question" and add help text for this field.
In AAQ step 3, display the Troubleshooting Info help text above the Share Data button, right after the field name.
In AAQ step 3, add the "tb_version" extra field.
In AAQ step 3, autofill the OS version field for Thunderbird.
Fix the issue with the Share Data button not working when editing a question.
Do not display the Common responses button when editing a question.
Improve accessibility.
Use "Windows 10/11" when there is not enough info to identify the exact Windows NT 10.0 version.